Finish multi instrument homing cleanup
