+++ AGENTS.md
@@ -44,7 +44,7 @@ Current high-level gantry code no longer applies a hidden Z sign flip. Working-v
 
 Use `docs/agent-index.md` for exact files/tests. Common entrypoints:
 
-- Gantry/motion/origin: `src/gantry/`, `setup/calibrate_deck_origin.py`, `setup/calibrate_multi_instrument_board.py`
+- Gantry/motion/origin: `src/gantry/`, `setup/calibrate_gantry.py`
 - Board/instrument offsets: `src/board/`
 - Deck/labware YAML: `src/deck/`, `configs/deck/`
 - Protocol YAML/setup/commands: `src/protocol_engine/`, `configs/protocol/`, `setup/validate_setup.py`
@@ -54,10 +54,9 @@ Use `docs/agent-index.md` for exact files/tests. Common entrypoints:
 
 ## Calibration Scripts
 
-- `setup/calibrate_deck_origin.py`: single-instrument deck-origin calibration.
-- `setup/calibrate_multi_instrument_board.py`: guided multi-instrument board calibration.
-  - CLI: `python setup/calibrate_multi_instrument_board.py --gantry <gantry.yaml>`
-  - Prompts the operator to explicitly choose the reference and lowest instruments by number; blank input is not accepted.
+- `setup/calibrate_gantry.py`: user-facing calibration script that requires a seed YAML and output gantry path, then chooses single- or multi-instrument calibration from instrument count.
+- `setup/calibrate_gantry.py`: only supported gantry calibration entrypoint. It contains both the single-instrument and multi-instrument flows and selects the correct path from instrument count in the seed YAML.
+  - Multi-instrument flow prompts the operator to explicitly choose the reference and lowest instruments by number; blank input is not accepted.
   - Starts guided jogging from the homed BRT pose; it does not make an automatic center move.
   - Sets XY with `G10 L20 P1 X0 Y0` only, then later sets Z to the calibration block height with `G10 L20 P1 Z<block_height>` using the lowest instrument.
   - Per-instrument calibration uses inverse `Board.move()` math against the same physical block point:
@@ -79,18 +78,11 @@ When the task is complete, either delete the temporary checkpoint after promotin
 
 ## Documentation Updates
 
-- **`calibrate_deck_origin.py`**: One-instrument deck-origin calibration utility for issue #87-style configs. Homes the machine at the normalized back-right-top homing corner, clears transient `G92` offsets, then prompts the operator to jog the reference TCP as far as appropriate toward the physical front-left XY origin and its lowest safe reachable Z. It sets only `G10 L20 P1 X0 Y0`, then assigns Z at the same pose. If the TCP touches true deck bottom, bottom mode sets `G10 L20 P1 Z0`. If the TCP cannot reach bottom, ruler-gap mode asks for the measured deck-to-TCP gap and sets `G10 L20 P1 Z<gap_mm>`. It then re-homes and reports measured physical maxima `(x_max, y_max, z_max)`. For one-instrument configs, use the lower-reach value as `working_volume.z_min`; for multi-instrument setups, prefer `calibrate_multi_instrument_board.py` so Z zero is defined by the lowest mounted instrument.
-    - **Guided usage**: `python setup/calibrate_deck_origin.py --gantry configs/gantry/cub_xl_asmi.yaml --instrument asmi`
-    - **Ruler gap for non-bottom-reaching TCP**: `python setup/calibrate_deck_origin.py --gantry <gantry.yaml> --z-reference-mode ruler-gap --tip-gap-mm 5 --instrument filmetrics`
-    - **Bottom contact**: `python setup/calibrate_deck_origin.py --gantry <gantry.yaml> --z-reference-mode bottom`
-    - **Dry run**: `python setup/calibrate_deck_origin.py --gantry <gantry.yaml> --dry-run`
-    - **Safety**: only use with deck-origin gantry configs whose X/Y working-volume minima are `0.0` and whose Z minimum is non-negative; pre-cutover or negative-space configs are rejected.
-- **`calibrate_multi_instrument_board.py`**: Guided multi-instrument calibration utility. Preconditions: GRBL `$3` axis directions and `$23` homing corner must be configured so `$H` homes to back-right-top and positive jog directions match the CubOS FLB deck frame. The flow homes, prompts the operator to explicitly pick the left-most/reference instrument by number, asks the operator to attach/jog it from the homed BRT pose to the front-left XY artifact, and sets only `G10 L20 P1 X0 Y0`. It intentionally does not make an automatic center move before the initial XY origining. It temporarily disables stale GRBL soft limits during the calibration jog flow, then re-homes for machine-derived X/Y bounds, moves to the measured XY center for calibration-block work, prompts the operator to explicitly pick and jog the lowest mounted instrument to the calibration block top, asks for the calibration block height, sets `G10 L20 P1 Z<block_height>`, re-homes for final Z max, moves back to the measured XY center, and records each instrument's `offset_x`, `offset_y`, and `depth` by having every instrument touch the same physical block point. The block's deck-frame X/Y/Z coordinates do not need to be known; place it near center where all instruments can reach and do not move it until calibration is complete.
-    - **Guided usage**: `python setup/calibrate_multi_instrument_board.py --gantry <gantry.yaml>`; the operator is prompted for the reference instrument and lowest instrument, and blank instrument selections are rejected.
-    - **Optional scripting flags**: `--reference-instrument` and `--lowest-instrument` can pre-fill prompts for tests or repeatable scripted runs.
-    - **Subset instruments**: repeat `--instrument <name>` to calibrate only selected instruments; otherwise all instruments in the gantry YAML are calibrated.
-    - **Output**: calibrated YAML is printed; use `--output-gantry <path>` or `--write-gantry-yaml` to write it after confirmation.
-    - **Safety**: this changes G54 WPos and may program GRBL soft limits; run `--dry-run` first and validate on hardware with slow jog steps, a calibration artifact/block, and clear E-stop access.
+- **`calibrate_gantry.py`**: Preferred user-facing calibration entrypoint. Loads a seed gantry YAML, requires a separate output path, preflights hardware risk and instrument count, then dispatches one-instrument configs to the internal deck-origin flow or multi-instrument configs to the internal board-calibration flow.
+    - **Guided usage**: `python setup/calibrate_gantry.py --seed configs/gantry/seeds/<seed>.yaml --output-gantry configs/gantry/<calibrated>.yaml`
+    - **Single instrument**: jog the tool to the calibration block at the front-left origin point; X/Y/Z are assigned at the same physical pose, with Z set to the block height.
+    - **Multi instrument**: pick the left-most/reference instrument and lowest instrument by number, use a calibration block, then record each instrument's `offset_x`, `offset_y`, and `depth` from the shared block point.
+    - **Safety**: this changes G54 WPos and may program GRBL soft limits; validate offline first and calibrate slowly with clear E-stop access.
 - **`hello_world.py`**: Interactive deck-origin jog test. Loads an explicit gantry YAML, homes without rewriting WPos, then lets you jog in the CubOS deck frame.
     - **Usage**: `python3 setup/hello_world.py --gantry configs/gantry/cub_xl_asmi.yaml`
     - **Controls**: Arrow keys (X/Y ±1mm), Z key (Z down 1mm), X key (Z up 1mm), Q (quit)
